Patch from Manuel Ruiz to make admin caommands easily callable from within code.
Refactor regarding creation of Context (new
InitialContext, getInitialContext in Ra....) into a Singleton class
(InitialContextBuilder) and if no environment is given or it is null, will try to
locate a ejbca.properties to extract from it the jndi properties
before using the default one.
Also make the ouputstream settable, with default beeing System.out.